test: should set config
