+++ src/tests/specs/mission-control/mission-control.cdn.spec.ts
@@ -6,7 +6,7 @@ import type { Principal } from '@dfinity/principal';
 import { type Actor, PocketIc } from '@hadronous/pic';
 import { beforeAll, describe, inject } from 'vitest';
 import { CONTROLLER_ERROR_MSG } from '../../constants/mission-control-tests.constants';
-import { testNotAllowedCdnMethods } from '../../utils/cdn-tests.utils';
+import { testControlledCdnMethods, testNotAllowedCdnMethods } from '../../utils/cdn-tests.utils';
 import { missionControlUserInitArgs } from '../../utils/mission-control-tests.utils';
 import { MISSION_CONTROL_WASM_PATH } from '../../utils/setup-tests.utils';
 
@@ -59,5 +59,13 @@ describe('Mission Control > Cdn', () => {
 		});
 
 		testNotAllowedCdnMethods({ actor: () => actor, errorMsg: CONTROLLER_ERROR_MSG });
+	});
+
+	describe('Admin', () => {
+		beforeAll(() => {
+			actor.setIdentity(controller);
+		});
+
+		testControlledCdnMethods({ actor: () => actor });
 	});
 });
